requirements.txt
numpy<2.0
PyQt5>=5.14.0,<5.15
opencv-python<4.6.0
mysql-connector-python
python-dotenv
whispercpp_kit
gTTS
setuptools
wheel
playsound==1.2.2
PyAudio
sounddevice
vosk
pycairo
PyGObject
openai
joblib
deepface
tf-keras
scikit-learn
markdown2
mediapipe
librosa